+++ libqtile/backend/base/__init__.py
@@ -1,0 +1,2 @@
+from libqtile.backend.base.core import Core  # noqa: F401
+from libqtile.backend.base.window import Internal, Static, Window  # noqa: F401
