+++ libqtile/backend/base/window.py
@@ -3,3 +3,11 @@ from abc import ABCMeta
 
 class Window(metaclass=ABCMeta):
     pass
+
+
+class Internal(metaclass=ABCMeta):
+    pass
+
+
+class Static(metaclass=ABCMeta):
+    pass
