+++ libqtile/hook.py
@@ -190,7 +190,7 @@ class Subscribe:
 
         **Arguments**
 
-            * ``window.Window`` object
+            * ``Window`` object
 
         Examples
         --------
@@ -214,7 +214,7 @@ class Subscribe:
 
         **Arguments**
 
-            * ``window.Window`` object of the managed window
+            * ``Window`` object of the managed window
         """
         return self._subscribe("client_managed", func)
 
@@ -223,7 +223,7 @@ class Subscribe:
 
         **Arguments**
 
-            * ``window.Window`` object of the killed window.
+            * ``Window`` object of the killed window.
         """
         return self._subscribe("client_killed", func)
 
@@ -232,7 +232,7 @@ class Subscribe:
 
         **Arguments**
 
-            * ``window.Window`` object of the new focus.
+            * ``Window`` object of the new focus.
         """
         return self._subscribe("client_focus", func)
 
@@ -241,7 +241,7 @@ class Subscribe:
 
         **Arguments**
 
-            * ``window.Window`` of window entered
+            * ``Window`` of window entered
         """
         return self._subscribe("client_mouse_enter", func)
 
@@ -250,7 +250,7 @@ class Subscribe:
 
         **Arguments**
 
-            * ``window.Window`` of client with updated name
+            * ``Window`` of client with updated name
         """
         return self._subscribe("client_name_updated", func)
 
@@ -259,7 +259,7 @@ class Subscribe:
 
         **Arguments**
 
-            * ``window.Window`` of client with hint change
+            * ``Window`` of client with hint change
         """
         return self._subscribe("client_urgent_hint_changed", func)
 
@@ -278,7 +278,7 @@ class Subscribe:
 
         **Arguments**
 
-            * ``window.Window`` of client with changed icon
+            * ``Window`` of client with changed icon
         """
         return self._subscribe("net_wm_icon_change", func)
