+++ libqtile/scratchpad.py
@@ -18,7 +18,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from libqtile import group, hook, window
+from libqtile import group, hook
+from libqtile.backend.x11 import xcbq
 
 
 class WindowVisibilityToggler:
@@ -98,7 +99,7 @@ class WindowVisibilityToggler:
             win = self.window
             # always set the floating state before changing group
             # to avoid disturbance of tiling layout
-            win._float_state = window.TOP
+            win._float_state = xcbq.TOP
             # add to group and bring it to front.
             win.togroup()
             win.cmd_bring_to_front()
