+++ libqtile/widget/clipboard.py
@@ -66,7 +66,7 @@ class Clipboard(base._TextBox):
         if owner_id in self.qtile.windows_map:
             owner = self.qtile.windows_map[owner_id].window
         else:
-            owner = xcbq.Window(self.qtile.conn, owner_id)
+            owner = xcbq.XWindow(self.qtile.conn, owner_id)
 
         owner_class = owner.get_wm_class()
         if owner_class:
